Replaced gluBuild2DMipmaps() calls with glGenerateMipmap().
This is the recommended way of making mipmaps these days. However, it causes TestTest.testSprites() to fail. This might be expected if filtering had changed slightly, but actually what is happening is two strange bars are appearing above the text, not in the alpha channel. The strange bars occur only on my IE Linux workstation, not my Mac. We should revisit this when moving to pure GL 3, and we have time to look into things - perhaps not using the fixed pipeline in TextPrimitive::renderSprites() will do the trick.