CI : Use GafferHQ/dependencies 3.0.0a2
This is only an alpha release in the sense that we will probably bump major version for some other dependencies (e.g. USD) before calling it 3.0.0. All the dependencies are from officially released (non-alpha) versions. The intention is to use 3.0.0 for Gaffer 0.60, but 3.0.0 is also a better approximation of the environment used for deployment at Image Engine.